Oops

+++ utilities_arrayops.py
@@ -355,7 +355,7 @@ def arrayops_concatenate(arrays, axis, dtype=None):
   if type(arrays[0]) is np.ndarray:
     return np.concatenate(arrays, axis=np.int32(axis), dtype = convert_to_npdtype(dtype))
   else:
-    return tf.concat(arrays, axis=tf.int32(axis), dtype = convert_to_tfdtype(dtype))
+    return tf.concat(arrays, axis=tf.cast(axis, tf.int32), dtype = convert_to_tfdtype(dtype))
 
 
 def arrayops_reshape(array, shape):
@@ -396,7 +396,7 @@ def arrayops_gather(array, indices, axis):
   if type(array) is np.ndarray:
     return np.take(array, indices, axis=np.int32(axis))
   else:
-    return tf.gather(array, indices, axis=tf.int32(axis))
+    return tf.gather(array, indices, axis=tf.cast(axis, tf.int32))
 
 
 def arrayops_gather_nd(array, indices, batch_dims=0):
